+++ src/CodeownerTeamsProvider.ts
@@ -2,6 +2,7 @@ import * as vscode from "vscode";
 import * as fs from "fs";
 import * as path from "path";
 import { getCodeownersTeams } from "./helpers/getCodeownersTeams";
+import { findCodeownersFile } from "./helpers/findCodeownersFile";
 
 const DIVIDER = "---";
 
@@ -30,20 +31,7 @@ export class CodeownerTeamsProvider
       return Promise.resolve([]);
     }
     
-    // Look for CODEOWNERS file in common locations
-    const possiblePaths = [
-      path.join(this.workspaceRoot, "CODEOWNERS"),
-      path.join(this.workspaceRoot, ".github", "CODEOWNERS"),
-      path.join(this.workspaceRoot, "docs", "CODEOWNERS")
-    ];
-    
-    let codeownersPath: string | undefined;
-    for (const possiblePath of possiblePaths) {
-      if (fs.existsSync(possiblePath)) {
-        codeownersPath = possiblePath;
-        break;
-      }
-    }
+    const codeownersPath = findCodeownersFile(this.workspaceRoot);
 
     if (!codeownersPath) {
       vscode.window.showInformationMessage("No CODEOWNERS file found");